requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-celery-beat==1.5.0
django-filter==2.2.0
django-rest-auth==0.9.5
django-timezone-field==3.1
djangorestframework==3.10.3
idna==2.8
importlib-metadata==0.23
itypes==1.1.0
Jinja2==2.10.3
kombu==4.6.5
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
oauthlib==3.1.0
Pygments==2.4.2
python-crontab==2.3.9
python-dateutil==2.8.0
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.10
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
vine==1.3.0
zipp==0.6.0
